requirements.txt
arrow==0.12.1
click==6.7
botometer==1.3
python-decouple==3.1
tweepy==3.6.0
